Similar treatment for StoreItems
Maybe consider settling on a naming convention
